requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pydantic
slowapi
freezegun
pytest-cov
